fix: balance update
