+++ src/renderer/entities/balance/model/balance-model.ts
@@ -1,10 +1,10 @@
 import { createEffect, createEvent, createStore, sample } from 'effector';
 import { throttle } from 'patronum';
-import keyBy from 'lodash/keyBy';
 
 import { Balance, ID } from '@shared/core';
 import { SAVE_TIMEOUT, BUFFER_DELAY } from '../lib/constants';
 import { storageService } from '@shared/api/storage';
+import { balanceUtils } from '../lib/balance-utils';
 
 const balancesUpdated = createEvent<Balance[]>();
 const balancesRemoved = createEvent<ID[]>();
@@ -23,27 +23,8 @@ const removeBalancesFx = createEffect(async (ids: ID[]): Promise<void> => {
 sample({
   clock: balancesUpdated,
   source: $balancesBuffer,
-  fn: (balances, newBalances) => {
-    const newBalancesMap = keyBy(newBalances, (b) => `${b.chainId}_${b.assetId}_${b.accountId}`);
-
-    const updatedBalances = balances.map((balance) => {
-      const { chainId, assetId, accountId } = balance;
-      const newBalance = newBalancesMap[`${chainId}_${assetId}_${accountId}`];
-
-      if (newBalance) {
-        balance.free = newBalance?.free || balance.free;
-        balance.frozen = newBalance?.frozen || balance.frozen;
-        balance.reserved = newBalance?.reserved || balance.reserved;
-        balance.locked = newBalance?.locked || balance.locked;
-
-        delete newBalancesMap[`${chainId}_${assetId}_${accountId}`];
-      }
-
-      return balance;
-    });
-
-    return updatedBalances.concat(Object.values(newBalancesMap));
-  },
+  filter: (_, newBalances) => newBalances.length > 0,
+  fn: balanceUtils.getMergeBalances,
   target: $balancesBuffer,
 });
